Added extensions for all storyboard and layout
